workshop1/requirements.txt
google-adk>=1.18.0
uvicorn
requests
